refactor: move webhook config into webhhook package
