+++ notify/webhook/webhook_test.go
@@ -32,7 +32,7 @@ import (
 	"github.com/stretchr/testify/require"
 	"gopkg.in/yaml.v2"
 
-	"github.com/prometheus/alertmanager/config"
+	amcommoncfg "github.com/prometheus/alertmanager/config/common"
 	"github.com/prometheus/alertmanager/notify"
 	"github.com/prometheus/alertmanager/notify/test"
 	"github.com/prometheus/alertmanager/types"
@@ -40,8 +40,8 @@ import (
 
 func TestWebhookRetry(t *testing.T) {
 	notifier, err := New(
-		&config.WebhookConfig{
-			URL:        config.SecretTemplateURL("http://example.com"),
+		&WebhookConfig{
+			URL:        amcommoncfg.SecretTemplateURL("http://example.com"),
 			HTTPConfig: &commoncfg.HTTPClientConfig{},
 		},
 		test.CreateTmpl(t),
@@ -109,8 +109,8 @@ func TestWebhookRedactedURL(t *testing.T) {
 
 	secret := "secret"
 	notifier, err := New(
-		&config.WebhookConfig{
-			URL:        config.SecretTemplateURL(u.String()),
+		&WebhookConfig{
+			URL:        amcommoncfg.SecretTemplateURL(u.String()),
 			HTTPConfig: &commoncfg.HTTPClientConfig{},
 		},
 		test.CreateTmpl(t),
@@ -131,7 +131,7 @@ func TestWebhookReadingURLFromFile(t *testing.T) {
 	require.NoError(t, err, "writing to temp file failed")
 
 	notifier, err := New(
-		&config.WebhookConfig{
+		&WebhookConfig{
 			URLFile:    f.Name(),
 			HTTPConfig: &commoncfg.HTTPClientConfig{},
 		},
@@ -191,8 +191,8 @@ func TestWebhookURLTemplating(t *testing.T) {
 			calledURL = "" // Reset for each test
 
 			notifier, err := New(
-				&config.WebhookConfig{
-					URL:        config.SecretTemplateURL(tc.url),
+				&WebhookConfig{
+					URL:        amcommoncfg.SecretTemplateURL(tc.url),
 					HTTPConfig: &commoncfg.HTTPClientConfig{},
 				},
 				test.CreateTmpl(t),
@@ -255,8 +255,8 @@ func TestWebhookDefaultPayload(t *testing.T) {
 	u, err := url.Parse("http://localhost")
 	require.NoError(t, err)
 
-	conf := &config.WebhookConfig{
-		URL:        config.SecretTemplateURL(u.String()),
+	conf := &WebhookConfig{
+		URL:        amcommoncfg.SecretTemplateURL(u.String()),
 		HTTPConfig: &commoncfg.HTTPClientConfig{},
 	}
 
@@ -312,8 +312,8 @@ func TestWebhookCustomPayloadMap(t *testing.T) {
 	u, err := url.Parse("http://localhost")
 	require.NoError(t, err)
 
-	conf := &config.WebhookConfig{
-		URL:        config.SecretTemplateURL(u.String()),
+	conf := &WebhookConfig{
+		URL:        amcommoncfg.SecretTemplateURL(u.String()),
 		HTTPConfig: &commoncfg.HTTPClientConfig{},
 		Payload: map[string]any{
 			"custom":       `some custom content`,
@@ -380,8 +380,8 @@ func TestWebhookCustomPayloadList(t *testing.T) {
 `), &payload)
 	require.NoError(t, err)
 
-	conf := &config.WebhookConfig{
-		URL:        config.SecretTemplateURL(u.String()),
+	conf := &WebhookConfig{
+		URL:        amcommoncfg.SecretTemplateURL(u.String()),
 		HTTPConfig: &commoncfg.HTTPClientConfig{},
 		Payload:    payload,
 	}
@@ -437,8 +437,8 @@ func TestWebhookCustomPayloadStringList(t *testing.T) {
 - foo: bar
 `
 
-	conf := &config.WebhookConfig{
-		URL:        config.SecretTemplateURL(u.String()),
+	conf := &WebhookConfig{
+		URL:        amcommoncfg.SecretTemplateURL(u.String()),
 		HTTPConfig: &commoncfg.HTTPClientConfig{},
 		Payload:    payload,
 	}
@@ -499,8 +499,8 @@ func TestWebhookCustomPayloadString(t *testing.T) {
 `
 	require.NoError(t, err)
 
-	conf := &config.WebhookConfig{
-		URL:        config.SecretTemplateURL(u.String()),
+	conf := &WebhookConfig{
+		URL:        amcommoncfg.SecretTemplateURL(u.String()),
 		HTTPConfig: &commoncfg.HTTPClientConfig{},
 		Payload:    payload,
 	}
